\name{lengthconvert-package}
\alias{lengthconvert-package}
\alias{lengthconvert}
\docType{package}
\title{
\packageTitle{lengthconvert}
}
\description{
\packageDescription{lengthconvert}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{lengthconvert}
\packageIndices{lengthconvert}
~~ An overview of how to use the package, including the ~~
~~ most important functions ~~
}
\author{
\packageAuthor{lengthconvert}

Maintainer: \packageMaintainer{lengthconvert}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, ~~
~~ from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
